Refine GitHub Enterprise Server support
This refines the tool's support for GitHub Enterprise Server,
targeting (for now) support for GHES 3.6 onwards.

When starting, the CLI checks the version of the API target
and errors if running an unsupported GHES version.

Auditors are passed `isGitHubEnterpriseServer` and
`gitHubEnterpriseServerVersion` arguments, and can decide how to
behave.

As part of this change, I have updated some existing auditors to
deal with GitHub Enterprise Server - for example, rulesets are
not available on GHES at all (and therefore the audit should no-op),
Actions variables are only available on GHES 3.8+ (and therefore
the audit should no-op on earlier versions) and some APIs exhibit
odd behaviour in GHES when Actions is disabled which must be
handled.