Suppress automatic warning from `octokit` about the tag protections API being deprecated (#198)
When a GitHub API is marked as deprecated, Octokit automatically
logs a deprecation warning when it is used. This is a bit annoying
and noisy. This suppresses that log by passing a custom logger
function to Octokit which eats that log line.

As a follow up to this change, I will make an issue to disable
the tag protections auditor on GitHub.com on the deprecation day,
August 30, 2024.

Fixes #190.